python葡萄叶病虫害图形识别/requirements.txt
flask
flask-cors
tensorflow
keras
efficientnet_pytorch
pillow
opencv-python
numpy
matplotlib
scikit-learn
pandas
tqdm
requests